Integrate the new `Repository` implementation
The new `Repository` implementation is integrated by wrapping it in a
new class `NodeRepository` that is mixed into the `Node` class. This
approach serves multiple purposes.

 1. The class serves as a backwards compatibility layer between the old
    repository interface, which would allow clients to set the mode for
    file handles for reading, as well as passing in text-like streams,
    and the new interface, which exclusively operates with bytes. The
    interface will decode the bytes when the caller is expecting normal
    strings in order to not break the interface.

 2. It allows to implement the concept of immutability which is a
    concept of the `Node` and is completely unknown to the `Repository`
    class.

 3. It nicely separates the file repository functionality into a
    separate file preventing the already huge `Node` implementation from
    getting even larger.

The downside of the latter is that the `NodeRepository` class contains
some potentially confusing logic. It accesses `Node` properties, but
from the code it is not clear where they come from. Since this class is
only meant to be used as a mixin for the `Node` class, I think that is
a sacrifice that is worth the benefits.