Add the repository mutating methods to the `Sealable` mixin.
This ensures that the repository is still mutable for sealable nodes as
long as they are not yet sealed. After sealing, the repository of even
these nodes becomes fully immutable.

Note that the implementation completely overrides the methods of the
`NodeRepositoryMixin`, because it needs to override the check in those
that will raise if the node is stored. An alternative would have been to
call the `super` from the `Sealable`, which due to the correct MRO,
would in fact the underlying method on the `NodeRepositoryMixin`, but it
would need to accept an argument, like for example `force`, to skip the
mutability check. This is not desirable, however, since those repository
methods are public methods of the `Node` interface and so any user will
be able to disable the mutability check for any node.

This solution does not suffer from that vulnerability but the downside
is that the implementation of the method from the `NodeRepositoryMixin`
needs to be copied and it needs to be kept in sync.