Add migrations for existing file repositories
The migration of an existing legacy file repository consists of:

 1. Make inventory of all files in the current file repository
 2. Write these files to the new backend (disk object store)
 3. Store the metadata of a node's repository contents, containing the
    virtual hierarchy to the corresponding `repository_metadata` column
    in the database.

The repository metadata will contain a hashkey for each file that was
written to the disk object store, which was generated by the latter and
that can be used to retrieve the content of the file.

The migration is performed in a single database migration, meaning that
everything is executed in a single transaction. Either the entire
migration succeeds or in case of an error, it is a no-op. This why the
migration will not delete the legacy file repository in the same
migration.

The advantage of this approach is that now there is no risk of data loss
and/or corruption. If the migration fails, all original data will be in
tact. The downside, however, is that the content of the file repository
is more or less duplicated. This means that the migration will require a
potentially significant amount of disk space that is at worst equal to
the size of the existing file repository. This should be the upper limit
since the disk object store will both deduplicate as well as compress
the content that is written.